# optimize get_block_generator()