--- a/benchmarks/utils.py
+++ b/benchmarks/utils.py
@@ -49,7 +49,9 @@ def rand_g2() -> G2Element:
 
 
 def rand_class_group_element() -> ClassgroupElement:
-    return ClassgroupElement(rand_bytes(100))
+    # TODO: address hint errors and remove ignores
+    #       error: Argument 1 to "ClassgroupElement" has incompatible type "bytes"; expected "bytes100"  [arg-type]
+    return ClassgroupElement(rand_bytes(100))  # type: ignore[arg-type]
 
 
 def rand_vdf() -> VDFInfo: