There are two parts to this optimization:

1. Perform the database lookup of all blocks simultaneously, instead of one at a time. This allows sqlite to more efficiently schedule the disk accesses.
2. Circumvent the normal `FullBlock` parser, and instead introduce a single-purpose parser that just picks out the generator program

The parser in (2) is not meant as a long term solution, but a stop-gap until we fix the parsing (and probably data structure) for `FullBlock` objects. That's not a trivial amount of work so this will probably be useful for a few versions.

Most of the code is a test that tests the parser on a wide range of `FullBlock` objects. Specifically, all combinations of optional fields are exercised. See `test_full_block_utilds.py`.

I've measured the timing of `get_generator()` in two scenarios, on SSD and HDD.

| system  | before | after | after / before |
| --- | --- | --- | --- |
| Ubuntu, SSD  | 1.767s | 0.220s | 12.5% |
| Ubuntu, HDD | 15.319s | 6.155s | 40.18 % |

As one would imagine, the HDD timing is dominated by the disk access.