@arvidn 2022-02-04
I'm working on adding tests for the new function on `BlockStore`. It also occurred to me that the v1 database case isn't covered by this optimization. I'm extending the patch to cover this case too.

I would also like to benchmark this patch against a v1 database, so I'm syncing one up right now.