@arvidn 2022-02-04
I have amended the benchmarks to also include a v1 database