@aqk 2022-02-04
`get_block_generator` is starting to get deeply nested